+++ docs/changelog.md
@@ -9,6 +9,8 @@ ITables ChangeLog
 
 **Fixed**
 - We have added an optional `columnControl` to `option.py`. Thanks to [kultura-luke](https://github.com/kultura-luke) for pointing out at this.
+- We have fixed one test on the streamlit example app ([#509](https://github.com/mwouts/itables/issues/509))
+- We have fixed the build step as well ([#513](https://github.com/mwouts/itables/issues/513))
 
 
 2.7.0 (2026-02-04)